moved all js files into tasks folder, deleted other code
